Hide the default text cursor when the mouse is idle for a certain duration
Previously, the code loaded and set system cursors for the default, changed, and default link cursors. In this commit, the code also loads and sets the system cursor for the default text cursor.

When the mouse is idle for a set amount of time, the code creates new copies of the default link and default text cursors and sets them as the system cursors. When the mouse is moved again, the code destroys the old copies and creates new copies using the changed cursor.

At the end of the program, all created cursors are destroyed.